requirements.txt
aiofiles==0.8.0
click==8.1.3
httptools==0.4.0
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
multidict==6.0.2
mysql-connector-python==8.0.30
protobuf==3.20.1
sanic==22.6.2
sanic-routing==22.3.0
ujson==5.4.0
uvloop==0.16.0
websockets==10.3
Werkzeug==2.2.2
zipp==3.8.1
